♻️ ReservationStateの状態の追加と、対応するNotifierのリファクタリング
