+++ lib/src/features/reserve/domain/reservation_state.dart
@@ -1,5 +1,6 @@
 enum ReservationState {
   noReservation,    // 予約なし
   hasReservation,   // 予約あり
+  hasPenalty,       // ペナルティあり
   error,            // エラー
 }
